requirements.txt
python-telegram-bot
pillow
requests
ollama
fake-useragent
github